so8t-mmllm/requirements.txt
# SO8T×マルチモーダルLLM（ローカル）依存関係
# RTX3060 12GB環境用

# コアMLライブラリ
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.41.2
accelerate>=0.24.0
bitsandbytes>=0.41.0

# マルチモーダル処理
qwen-vl-utils>=0.0.1
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0

# データ処理
datasets>=2.14.0
sentencepiece>=0.1.99
tokenizers>=0.15.0

# 学習・最適化
peft>=0.6.0
trl>=0.7.0
wandb>=0.16.0
tensorboard>=2.14.0

# データベース
sqlite3

# ユーティリティ
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 開発・テスト
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# 推論・配布
llama-cpp-python>=0.2.0
vllm>=0.2.0
